Hero section functionality and responsiveness
Arranged hero section in index.html with both new and existing CSS classes to make the hero section more responsive and functional.  Created a container to hold the text and button, so they could be visible when the screen shrinks and they move over the darker city skyline.  Changed font-size in media queries for hero text so the text stays within the container.  Positioned the logo on the left of the text, and gave it a max width so that it does not become oblong as it grows, and a min width so it does not get smaller than what is needed to complement the text.